Fixes blurry text in the sidebar
On Google Chrome, Windows 10:

When using the `position: sticky;` style, there's currently a bug that makes the text blurry in the sticky container if the `top` is not a round number of pixels.
This stylesheet uses relative em units which causes the top to be `10.4px`. When setting the top to 10px, the text is sharp again!

I've also tested this on Firefox, Edge and IE, where it's not an issue. So far I've only found [https://bugs.webkit.org/show_bug.cgi?id=159428](this webkit issue) that even begins to explain it.