more obvious and ever-present callout when not on current version
currently the only thing that indicates a user is not on the current version on the documentation is the version dropdown in the upper right corner, and a callout box, both of which are only visible at the top of the page.

it would be great to have some ever-present visual to indicate to the user they ARE NOT on the current documentation. I know I've had times I've forgotten I had old documentation up.

open to other designs, but this is what I started with.